+++ app/Http/Controllers/ProductInfoApiController.php
@@ -10,30 +10,33 @@ use Illuminate\Support\Facades\Auth;
 
 class ProductInfoApiController extends Controller
 {
+
     public function index($hash)
     {
         $user = Auth::user();
-        if (!$user->tokenCan('read')) return response('Forbidden', 403);
+        if (!$user->tokenCan('read') || !$user->has_pro_plan() || $user->is_over_api_limit()) return response('Forbidden', 403);
         $product = Product::with("product_infos", "product_infos.language", "owner")->where('hash', $hash)->first();
         if (!$product) return response('Not Found', 404);
         if ($user->id != $product->user_id) return response('Forbidden', 403);
+        $user->call_api();
         return response()->json($product);
     }
 
     public function show($hash)
     {
         $user = Auth::user();
-        if (!$user->tokenCan('read')) return response('Forbidden', 403);
+        if (!$user->tokenCan('read') || !$user->has_pro_plan() || $user->is_over_api_limit()) return response('Forbidden', 403);
         $info = ProductInfo::with("language", "owner")->where('hash', $hash)->first();
         if (!$info) return response('Not Found', 404);
         if ($user->id != $info->user_id) return response('Forbidden', 403);
+        $user->call_api();
         return response()->json($info);
     }
 
     public function store(Request $request, $hash)
     {
         $user = Auth::user();
-        if (!$user->tokenCan('create')) return response('Forbidden', 403);
+        if (!$user->tokenCan('create') || !$user->has_pro_plan() || $user->is_over_api_limit()) return response('Forbidden', 403);
         $product = Product::where('hash', $hash)->first();
         if (!$product) return response('Not Found', 404);
         if ($user->id != $product->user_id) return response('Forbidden', 403);
@@ -60,12 +63,14 @@ class ProductInfoApiController extends Controller
             }
         }
         $info = ProductInfo::with("sections")->findOrFail($info->id);
+        $user->call_api();
         return response()->json($info);
     }
 
     public function update(Request $request, $hash, $hash2)
     {
         $user = Auth::user();
+        if(!$user->has_pro_plan() || $user->is_over_api_limit()) return response('Forbidden', 403);
         $product = Product::where('hash', $hash)->first();
         if (!$product) return response('Not Found', 404);
         if ($user->cannot('update', $product)) return response('Forbidden', 403);
@@ -97,12 +102,14 @@ class ProductInfoApiController extends Controller
                 $sec_new->save();
             }
         }
+        $user->call_api();
         return response()->json($info->with('sections'));
     }
 
     public function destroy($hash)
     {
         $user = Auth::user();
+        if(!$user->has_pro_plan() || $user->is_over_api_limit()) return response('Forbidden', 403);
         $info = ProductInfo::where('hash', $hash)->first();
         if (empty($info)) return response('Not Found', 404);
         if ($user->cannot('delete', $info)) return response('Forbidden', 403);
@@ -113,10 +120,12 @@ class ProductInfoApiController extends Controller
     public function destroy_section($hash)
     {
         $user = Auth::user();
+        if(!$user->has_pro_plan() || $user->is_over_api_limit()) return response('Forbidden', 403);
         $section = ProductInfoSection::where('hash', $hash)->first();
         if ($section->isEmpty()) return response('Not Found', 404);
         if ($user->cannot('delete', $section)) return response('Forbidden', 403);
         $section->delete();
+        $user->call_api();
         return response('Success');
     }
 }
